refactor: extract useVisibleChunkIndices hook to eliminate Rule of Three
The identical chunk visibility selector was copy-pasted across Channel,
TimeScale, and SpectrogramChannel. Extract into a shared hook with
memoized return value for stable effect dependencies.

Also removes an unnecessary variable alias in PlaylistVisualization
and commits CLAUDE.md pattern updates from the useDynamicTracks work.
