+++ packages/ui-components/src/contexts/index.tsx
@@ -3,7 +3,7 @@ import { usePlaylistInfo, PlaylistInfoContext } from './PlaylistInfo';
 import { useTheme } from './Theme';
 import { useTrackControls, TrackControlsContext } from './TrackControls';
 import { PlayoutProvider, usePlayoutStatus, usePlayoutStatusUpdate } from './Playout';
-import { useScrollViewport, useScrollViewportSelector, ScrollViewportProvider } from './ScrollViewport';
+import { useScrollViewport, useScrollViewportSelector, useVisibleChunkIndices, ScrollViewportProvider } from './ScrollViewport';
 export type { ScrollViewport } from './ScrollViewport';
 
 export {
@@ -19,5 +19,6 @@ export {
   usePlayoutStatusUpdate,
   useScrollViewport,
   useScrollViewportSelector,
+  useVisibleChunkIndices,
   ScrollViewportProvider,
 };
